Feature/create commerce (#6)
* changes

* create user

* creando branch

* campaing

* refactor dtos

* commerce logic

* fix

+++ src/domain/entities/domain/branch.go
@@ -3,9 +3,9 @@ package domain
 import "time"
 
 type Branch struct {
-	ID         uint       `gorm:"primaryKey"`
+	ID         uint64     `gorm:"primaryKey"`
 	Name       string     `json:"name" validate:"required"`
-	CommerceID uint       `json:"commerce_id" validate:"required"`
+	CommerceID uint64     `json:"commerce_id" validate:"required"`
 	Address    string     `json:"address"`
 	StartDate  time.Time  `json:"start_date"`
 	EndDate    *time.Time `json:"end_date,omitempty"`
